refactor(llm): rename cache_markers_pi -> cache_markers (M6.3)
Drop the _pi suffix from the cubepi cache-marker module and its unit test.
Class CubeboxCacheMarkerPolicy already had no Pi suffix; only import sites
updated (run_manager.py + test file).
