+++ backend/cubebox/streams/run_manager.py
@@ -475,7 +475,7 @@ class RunManager:
         from cubebox.agents.graph import create_cubebox_agent
         from cubebox.agents.stream import convert_agent_event_to_sse
         from cubebox.db.engine import async_session_maker
-        from cubebox.llm.cache_markers_pi import CubeboxCacheMarkerPolicy
+        from cubebox.llm.cache_markers import CubeboxCacheMarkerPolicy
         from cubebox.llm.factory import LLMFactory
 
         try:
